requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.1.7
django-chartjs==2.3.0
django-cors-headers==4.0.0
django-crispy-forms==2.0
django-extensions==3.2.1
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
numpy==1.24.3
pandas==1.5.3
pipreqs==0.4.13
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.3
yarg==0.1.9
